backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
django-environ==0.11.2

# Channels & WebSocket
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Database
psycopg2-binary==2.9.10

# Celery
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.2.1

# Image handling
Pillow==11.1.0

# Utilities
python-dateutil==2.9.0
whitenoise==6.8.2
gunicorn==23.0.0

# Stripe payments
stripe==11.4.1
